Bump to 1.1.4
Bumping to 1.1.4 to unify the versions specified in pathsearch.py and
pyproject.toml. This is needed because I forgot to update the version in
pathsearch.py for 1.1.3, so if you install 1.1.3 and run it with -V,
it will tell you it's version 1.1.2. I cannot just change the version in
pathsearch.py and reupload it to PyPI as 1.1.3, so this is a workaround.